assume qint64 and ssize_t are the same in QNativeSocketEnginePrivate::nativeWrite()

+++ src/network/socket/qnativesocketengine_unix.cpp
@@ -907,7 +907,7 @@ qint64 QNativeSocketEnginePrivate::nativeWrite(const char *data, qint64 len)
 {
     Q_Q(QNativeSocketEngine);
 
-    ssize_t writtenBytes = qt_safe_write_nosignal(socketDescriptor, data, len);
+    qint64 writtenBytes = qt_safe_write_nosignal(socketDescriptor, data, len);
 
     if (writtenBytes < 0) {
         switch (errno) {
@@ -934,7 +934,7 @@ qint64 QNativeSocketEnginePrivate::nativeWrite(const char *data, qint64 len)
                                 (int) len).data(), len, (int) writtenBytes);
 #endif
 
-    return qint64(writtenBytes);
+    return writtenBytes;
 }
 /*
 */
